requirements.txt
streamlit
pandas
gspread
oauth2client
google-generativeai>=0.8.0
pillow